I'm not so sure about having BandSelector only on CM7 side.
