+++ cmake/platforms/rpi/rpi.cmake
@@ -157,6 +157,26 @@ set(EVENT_SOURCES
         src/core/radio/transmitter/events/qt/TransmitterIqEvent.cpp
 )
 
+set (RADIO_SOURCES
+        ${RADIO_BASE_SOURCES}
+        ${RADIO_MASTER_SOURCES}
+        ${RADIO_ENGINE_SOURCES}
+        src/core/radio/Radio.cpp
+        src/core/radio/Radio.h
+)
+
+set (CORE_SOURCES
+        ${CONFIG_SOURCES}
+        ${SETTINGS_SOURCES}
+        ${DSP_SOURCES}
+        ${IO_SOURCES}
+        ${RADIO_SOURCES}
+        ${EVENT_SOURCES}
+        ${UTIL_SOURCES}
+        ${USB_API_SOURCES}
+        src/core/SampleTypes.h
+)
+
 if(USE_GUI)
     set(QT_GUI_SOURCES
             src/app/qt/common/QtChartTheme.h
