+++ src/core/config-settings/settings/RadioSettings.h
@@ -14,7 +14,7 @@
 #include <vector>
 
 #include "bands/BandSettings.h"
-#include "bands/BandSelector.h"
+#include "bands/ActiveBandSettings.h"
 #include "base/SettingsException.h"
 
 
@@ -56,19 +56,20 @@ public:
     return *this;
   }
 
-  BandSelector& getBandSelector() { return m_bandSelector; }
+  ActiveBandSettings& getActiveBandSettings() { return m_activeBandSettings; }
 
   [[nodiscard]] bool getPtt() const { return m_ptt(); }
   // [[nodiscard]] const std::string& getBandName() const { return m_bandName(); }
   [[nodiscard]] const ReceiverSettings& getRxSettings() const { return m_rxSettings; }
   [[nodiscard]] const TransmitterSettings& getTxSettings() const { return m_txSettings; }
 
-  static void setCentreFrequencyDeltas(int32_t fine, int32_t coarse);
-  static BandSettings* getBandSettings(const std::string& bandName);
-  static BandSettings* getFocusBandSettings();
-  static const std::string& getFocusBandName() { return m_bandSelector.getFocusBandName(); }
+  void setCentreFrequencyDeltas(int32_t fine, int32_t coarse);
+  BandSettings* getBandSettings(const std::string& bandName);
+  BandSettings* getFocusBandSettings();
+  const BandSettings* getFocusBandSettings() const;
+  const std::string& getFocusBandName() const { return m_activeBandSettings.getFocusBandName(); }
 
-  static const Bands& getBands() { return m_bandSelector.getAllBands(); }
+  const Bands& getBands() { return m_activeBandSettings.getAllBands(); }
 
   void applyRfSettings(const RfSettings& settings, bool onlyChanged = false);
   void applyIfSettings(const IfSettings& settings);
@@ -111,5 +112,5 @@ protected:
   // Setting<std::string, BAND, RadioSettings> m_bandName;
   ReceiverSettings m_rxSettings;
   TransmitterSettings m_txSettings;
-  static BandSelector m_bandSelector;
+  ActiveBandSettings m_activeBandSettings;
 };
