+++ src/core/radio/RadioFrontEnd.cpp
@@ -1,0 +1,46 @@
+//
+// Created by murray on 23/3/26.
+//
+
+#include "RadioFrontEnd.h"
+
+RadioFrontEnd::RadioFrontEnd(EventTarget *pEventTarget) :
+  RadioBase(pEventTarget)
+{
+
+}
+
+void
+RadioFrontEnd::configure(const RadioConfig* pConfig)
+{
+}
+
+void
+RadioFrontEnd::start()
+{
+
+}
+
+void
+RadioFrontEnd::stop()
+{
+
+}
+
+void
+RadioFrontEnd::applySettings(const RadioSettings& settings)
+{
+
+}
+
+void
+RadioFrontEnd::applySettingUpdate(SettingUpdate& update)
+{
+
+}
+
+void
+RadioFrontEnd::notifyUpdate(const SettingUpdate& update)
+{
+
+}
